Refactor: split client.go and model.go into resource-specific files
- Extract Server, Switch, DNS, ELB, GSLB, DB, Disk resources into separate files
- Move render*Detail functions from model.go to render.go
- Keep common definitions (ResourceType, SakuraClient) in client.go
- Improve code maintainability for future resource implementations
